clean up SparkMicroBatchStreamTest
